+++ app/models/idea.rb
@@ -1,7 +1,7 @@
 class Idea < ActiveRecord::Base
+  include Likeable
 	belongs_to :user
 	has_many :comments
-  has_many :likes, as: :likeable
 	mount_uploader :picture, PictureUploader
 
   validates :name, presence: true, length: {
@@ -15,13 +15,4 @@ class Idea < ActiveRecord::Base
   }
   validates :picture, presence: true
   validates :user, presence: true
-
-  def liked_by?(user)
-    liked_by(user).present?
-  end
-
-  def liked_by(user_id)
-    user_id = user_id.id if user_id.is_a?(User)
-    likes.find_by(user_id: user_id)
-  end
 end
