+++ app/views/ideas/show.html.erb
@@ -20,6 +20,7 @@
         <strong><%= comment.user_name %></strong>
         <br/>
         <p><%= comment.body %></p>
+        <p><%= render 'likes/like_button', likeable: comment %></p>
         <p><%= link_to 'Delete', comment_path(comment), method: :delete, data: { confirm: 'Are you sure?' } %></p>
       </div>
     <% end %>
