+++ app/views/likes/_create_button.html.erb
@@ -1,3 +1,1 @@
-<span class="likeable-link">
-  <%= link_to 'Like', likes_path(like: {likeable_id: likeable.id, likeable_type: likeable.class.to_s}), method: :post, remote: true %>
-</span>
+<%= link_to 'Like', likes_path(like: {likeable_id: likeable.id, likeable_type: likeable.class.to_s}), method: :post, remote: true %>
