+++ app/views/likes/_destroy_button.html.erb
@@ -1,3 +1,1 @@
-<span class="likeable-link">
-  <%= link_to 'Unlike', likeable.liked_by(current_user), method: :delete, remote: true %>
-</span>
+<%= link_to 'Unlike', likeable.liked_by(current_user), method: :delete, remote: true %>
